remove spurious sorting of Report in error_report capture
Report as described in section 2.3 of
http://www.erlang.org/doc/apps/sasl/error_logging.html is a list of
the crashed process' info and any neighbors' infos.
Example code that creates such a report can be found in proc_lib, e.g.
the ultimate try/catch of proc_lib:spawn'd process:
https://github.com/erlang/otp/blob/maint/lib/stdlib/src/proc_lib.erl#L492-L496

Raven's error_logger handler was sorting this list, which reversed
the order, and misconstrued the crashed process' info as the neighbors'.