Support Sentry users for general error macros
The ?failed and ?exception macros get picked up by parse_message already
and allow attaching arbitray extra data to be sent to Sentry.

This makes it so that attaching `user` as extra data is sent as a proper
raven user, allowing Sentry to do smart things like count the number of
affected users and searching for all issues related to this user.

Sentry expects certain keys to work, from the docs:

id:
    The unique ID of the user.
email:
    The email address of the user.
ip_address:
    The IP of the user.
username:
    The username of the user.

All other keys are stored as extra information but not specifically
processed by Sentry.

To make this easier, if you don't provide a proplist/eon object, it will
default to using the `user` key as `id`.